requirements.txt
mysql-connector-python==9.1.0
chromadb-client==1.0.13
sentence-transformers==3.2.0
requests==2.32.3
python-dotenv==1.1.0
openai==1.88.0